PrometheusHealthCheck gets its own feature
