+++ src/Harald.WebApi/Features/PrometheusHealthCheck/Configuration/Middleware.cs
@@ -1,0 +1,20 @@
+using Microsoft.AspNetCore.Builder;
+using Microsoft.AspNetCore.Diagnostics.HealthChecks;
+
+namespace Harald.WebApi.Features.PrometheusHealthCheck.Configuration
+{
+    public static class Middleware
+    {
+        public static IApplicationBuilder UsePrometheusHealthCheck(
+            this IApplicationBuilder app
+        )
+        {
+            app.UseHealthChecks("/healthz", new HealthCheckOptions
+            {
+                ResponseWriter = PrometheusHealthProbe.WriteHealthResponseAsync
+            });
+
+            return app;
+        }
+    }
+}
