+++ src/Harald.WebApi/Startup.cs
@@ -1,29 +1,23 @@
 ﻿using System;
 using System.Net;
-using System.Threading;
-using System.Threading.Tasks;
 using Harald.Infrastructure.Slack;
-using Harald.WebApi.Application;
 using Harald.WebApi.Application.EventHandlers;
 using Harald.WebApi.Domain;
 using Harald.WebApi.Domain.Events;
 using Harald.WebApi.Features.Connections.Configuration;
 using Harald.WebApi.Features.Metrics.Configuration;
+using Harald.WebApi.Features.PrometheusHealthCheck.Configuration;
 using Harald.WebApi.Infrastructure.Messaging;
 using Harald.WebApi.Infrastructure.Persistence;
 using Harald.WebApi.Infrastructure.Serialization;
 using Harald.WebApi.Infrastructure.Services;
 using Microsoft.AspNetCore.Builder;
-using Microsoft.AspNetCore.Diagnostics.HealthChecks;
 using Microsoft.AspNetCore.Hosting;
-using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Diagnostics.HealthChecks;
-using Microsoft.Extensions.Hosting;
-using Prometheus;
 
 namespace Harald.WebApi
 {
@@ -151,59 +145,7 @@ namespace Harald.WebApi
 
             app.UseMvc();
 
-            app.UseHealthChecks("/healthz", new HealthCheckOptions
-            {
-                ResponseWriter = MyPrometheusStuff.WriteResponseAsync
-            });
-        }
-    }
-
-
-    public static class MyPrometheusStuff
-    {
-        private const string HealthCheckLabelServiceName = "service";
-        private const string HealthCheckLabelStatusName = "status";
-
-        private static readonly Gauge HealthChecksDuration;
-        private static readonly Gauge HealthChecksResult;
-
-        static MyPrometheusStuff()
-        {
-            HealthChecksResult = Metrics.CreateGauge("healthcheck",
-                "Shows health check status (status=unhealthy|degraded|healthy) 1 for triggered, otherwise 0",
-                new GaugeConfiguration
-                {
-                    LabelNames = new[] {HealthCheckLabelServiceName, HealthCheckLabelStatusName},
-                    SuppressInitialValue = false
-                });
-
-            HealthChecksDuration = Metrics.CreateGauge("healthcheck_duration_seconds",
-                "Shows duration of the health check execution in seconds",
-                new GaugeConfiguration
-                {
-                    LabelNames = new[] {HealthCheckLabelServiceName},
-                    SuppressInitialValue = false
-                });
-        }
-
-        public static Task WriteResponseAsync(HttpContext httpContext, HealthReport healthReport)
-        {
-            UpdateMetrics(healthReport);
-
-            httpContext.Response.ContentType = "text/plain";
-            return httpContext.Response.WriteAsync(healthReport.Status.ToString());
-        }
-
-        private static void UpdateMetrics(HealthReport report)
-        {
-            foreach (var (key, value) in report.Entries)
-            {
-                HealthChecksResult.Labels(key, "healthy").Set(value.Status == HealthStatus.Healthy ? 1 : 0);
-                HealthChecksResult.Labels(key, "unhealthy").Set(value.Status == HealthStatus.Unhealthy ? 1 : 0);
-                HealthChecksResult.Labels(key, "degraded").Set(value.Status == HealthStatus.Degraded ? 1 : 0);
-
-                HealthChecksDuration.Labels(key).Set(value.Duration.TotalSeconds);
-            }
+            app.UsePrometheusHealthCheck();
         }
     }
 }
